Added src/types/ and src/utils/, moved types definitions into types and filtering/sorting logic into utils
